build(proto): split content graph api into its own file
