+++ frontend/shared/src/client/.generated/documents/v1alpha/content_graph_connectweb.ts
@@ -1,0 +1,28 @@
+// @generated by protoc-gen-connect-web v0.6.0 with parameter "target=ts,import_extension=none"
+// @generated from file documents/v1alpha/content_graph.proto (package com.mintter.documents.v1alpha, syntax proto3)
+/* eslint-disable */
+// @ts-nocheck
+
+import { ListCitationsRequest, ListCitationsResponse } from "./content_graph_pb";
+import { MethodKind } from "@bufbuild/protobuf";
+
+/**
+ * Content graph service provides access to citations (backlinks).
+ *
+ * @generated from service com.mintter.documents.v1alpha.ContentGraph
+ */
+export const ContentGraph = {
+  typeName: "com.mintter.documents.v1alpha.ContentGraph",
+  methods: {
+    /**
+     * @generated from rpc com.mintter.documents.v1alpha.ContentGraph.ListCitations
+     */
+    listCitations: {
+      name: "ListCitations",
+      I: ListCitationsRequest,
+      O: ListCitationsResponse,
+      kind: MethodKind.Unary,
+    },
+  }
+} as const;
+
