+++ proto/documents/v1alpha/content_graph.proto
@@ -1,0 +1,49 @@
+syntax = "proto3";
+
+package com.mintter.documents.v1alpha;
+
+option go_package = "mintter/backend/genproto/documents/v1alpha;documents";
+
+// Content graph service provides access to citations (backlinks).
+service ContentGraph {
+  rpc ListCitations(ListCitationsRequest) returns (ListCitationsResponse);
+}
+
+message ListCitationsRequest {
+  // Required. Document ID for which citations need to be retrieved.
+  string document_id = 1;
+
+  // Optional. Depth can be used to request transitive closure of citations.
+  // For example depth=1 will return not only citations of the requested document_id
+  // but also citations of those direct citations. The default is depth=0 and will only
+  // return direct citations.
+  int32 depth = 2;
+}
+
+// Response with citations.
+message ListCitationsResponse {
+  // List of links that point to the requested document, recursively, according to the requested depth.
+  repeated Link links = 1;
+}
+
+// Description of a link inside a document.
+message Link {
+  // Required. Describes where link originates from.
+  LinkNode source = 1;
+
+  // Required. Describes where link points to.
+  // Here the block_id is optional, because the whole document can be linked.
+  LinkNode target = 2;
+}
+
+// Describes "sides" of a Link.
+message LinkNode {
+  // ID of the document on one side of a Link.
+  string document_id = 1;
+
+  // Version of the document.
+  string version = 2;
+
+  // ID of the block inside the document.
+  string block_id = 3;
+}
