+++ proto/documents/v1alpha/go.gensum
@@ -1,2 +1,2 @@
-srcs: ce6db46e921fef408619c1271f0817b2
-outs: 1b29db0abc14bd4fe3e1f37f95050350
+srcs: 94377bafbc5df7334dd39a098c433f66
+outs: dd03cccc9b665babefad529f8098087d
