+++ proto/documents/v1alpha/js.gensum
@@ -1,2 +1,2 @@
-srcs: ce6db46e921fef408619c1271f0817b2
-outs: 7b9f891cff00b3f35f33768286a82dbd
+srcs: 94377bafbc5df7334dd39a098c433f66
+outs: 0316393af3ec7482256e5ffd28790ea9
